Start splitting licence import
Currently, we have licence import setup to do one licence at a time. But this is purely because we have to deal with 30-ish licences differently because they don't have an effective start date.

To squeeze a little more performance out of the process, and reduce our hits to the DB by more than 73 thousand (!) we're splitting the licence import.

This will now handle just those with no start date. For the rest we'll use the same process as for versions, purposes etc, a single UPSERT query we can apply to all licences.
